refactor: reorganize ui state stores
