+++ src/ui/views/DesktopSmallSwap/hooks/useBatchSwapTask.ts
@@ -2,7 +2,7 @@ import { Account } from '@/background/service/preference';
 import { DEX, KEYRING_TYPE } from '@/constant';
 import { useMiniSigner } from '@/ui/hooks/useSigner';
 import { useRabbySelector } from '@/ui/store';
-import { useSwapStore } from '@/ui/stores/swap';
+import { useSwapStore } from '@/ui/state/swap';
 import { formatAmount, useWallet, WalletControllerType } from '@/ui/utils';
 import { waitForTxCompleted } from '@/ui/utils/transaction';
 import { useGasAccountSign } from '@/ui/views/GasAccount/hooks';
